cofiguring front end to make room for log in page
